Fix warning in spam_report
spam_report.c:239:25: warning: call to '_curl_easy_setopt_err_long'
declared with attribute warning: curl_easy_setopt expects a long
argument

+++ src/plugins/spam_report/spam_report.c
@@ -236,7 +236,7 @@ static void report_spam(gint id, ReportInterface *intf, MsgInfo *msginfo, gchar 
 			curl_easy_setopt(curl, CURLOPT_URL, intf->url);
 			curl_easy_setopt(curl, CURLOPT_POSTFIELDS, reqbody);
 			curl_easy_setopt(curl, CURLOPT_USERPWD, auth);
-			curl_easy_setopt(curl, CURLOPT_TIMEOUT, prefs_common_get_prefs()->io_timeout_secs);
+			curl_easy_setopt(curl, CURLOPT_TIMEOUT, (long)(prefs_common_get_prefs()->io_timeout_secs));
 			curl_easy_setopt(curl, CURLOPT_USERAGENT,
                 		SPAM_REPORT_USERAGENT "(" PLUGINS_URI ")");
 #ifdef G_OS_WIN32
